pyproject.toml
[tool.poetry]
name = "jobfetcher"
version = "0.1.0"
description = "Job search aggregation agent for scraping LinkedIn"
authors = ["JobFetcher Team"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/jobfetcher/jobfetcher"
repository = "https://github.com/jobfetcher/jobfetcher"
keywords = ["job", "scraper", "linkedin"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.0"
beautifulsoup4 = "^4.12.0"
lxml = "^5.0.0"
playwright = "^1.40.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.5.0"
sqlalchemy = "^2.0.0"
apscheduler = "^3.10.0"
python-dotenv = "^1.0.0"
tqdm = "^4.66.0"
python-dateutil = "^2.8.0"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
jobfetcher = "jobfetcher.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
